% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normalize_data.R
\name{normalize_data}
\alias{normalize_data}
\title{Normalize Data Using Z-Score Standardization}
\usage{
normalize_data(mat)
}
\arguments{
\item{mat}{A numeric matrix. The function will normalize each column of the matrix.}
}
\value{
A numeric matrix with the same dimensions as `mat`, where each column has
been standardized to have zero mean and unit variance.
}
\description{
This function performs column-wise normalization on a numeric matrix using
Z-score standardization. Each column will be transformed to have a mean of 0
and a standard deviation of 1.
}
\details{
This function manually computes the mean and standard deviation for each
column of the matrix, then applies Z-score normalization:
\deqn{z = \frac{x - \bar{x}}{s}}
where \eqn{\bar{x}} is the column mean and \eqn{s} is the column standard deviation.
}
\examples{
mat <- matrix(c(1, 2, 3, 4, 5, 6), nrow = 3)
normalize_data(mat)

}
